[Feat/#179] :recycle: Refactor: 패키지 구조 리팩토링
