--- chatty-be/src/main/java/com/chatty/chatty/aws/entity/Question.java
@@ -1,19 +1,0 @@
-package com.chatty.chatty.aws.entity;
-
-import java.util.List;
-
-public record Question(
-
-        String id,
-
-        Integer question_number,
-
-        String type,
-
-        String question,
-
-        List<String> options,
-
-        String answer
-) {
-}
