+++ chatty-be/src/main/java/com/chatty/chatty/game/config/dynamodb/DynamoDBConfig.java
@@ -1,0 +1,26 @@
+package com.chatty.chatty.game.config.dynamodb;
+
+import com.amazonaws.auth.AWSStaticCredentialsProvider;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
+import com.chatty.chatty.config.AWSKey;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@RequiredArgsConstructor
+public class DynamoDBConfig {
+
+    private final AWSKey awsKey;
+
+    @Bean
+    public AmazonDynamoDB amazonDynamoDB() {
+        return AmazonDynamoDBClientBuilder.standard()
+                .withCredentials(
+                        new AWSStaticCredentialsProvider(new BasicAWSCredentials(awsKey.getAccessKey(), awsKey.getSecretKey())))
+                .withRegion(awsKey.getRegion())
+                .build();
+    }
+}
