+++ chatty-be/src/main/java/com/chatty/chatty/game/config/model/RestClientConfig.java
@@ -1,0 +1,16 @@
+package com.chatty.chatty.game.config.model;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestClient;
+
+@Configuration
+public class RestClientConfig {
+
+    @Bean
+    public RestClient restClient() {
+        return RestClient.builder()
+                .baseUrl("http://localhost:8000")
+                .build();
+    }
+}
