+++ chatty-be/src/main/java/com/chatty/chatty/game/controller/dto/ChatRequest.java
@@ -1,0 +1,11 @@
+package com.chatty.chatty.game.controller.dto;
+
+public record ChatRequest(
+        ChatType chatType,
+
+        Long roomId,
+
+        String message
+) {
+
+}
