+++ chatty-be/src/main/java/com/chatty/chatty/game/controller/dto/ChatResponse.java
@@ -1,0 +1,19 @@
+package com.chatty.chatty.game.controller.dto;
+
+import java.time.LocalDateTime;
+import lombok.Builder;
+
+@Builder
+public record ChatResponse(
+        ChatType chatType,
+
+        Long roomId,
+
+        Long userId,
+
+        String message,
+
+        LocalDateTime time
+) {
+
+}
