+++ chatty-be/src/main/java/com/chatty/chatty/game/controller/dto/ChatType.java
@@ -1,0 +1,7 @@
+package com.chatty.chatty.game.controller.dto;
+
+public enum ChatType {
+
+    TEXT,
+    EMOJI
+}
