+++ chatty-be/src/main/java/com/chatty/chatty/game/controller/dto/PlayersStatusDTO.java
@@ -1,0 +1,15 @@
+package com.chatty.chatty.game.controller.dto;
+
+import com.chatty.chatty.player.domain.PlayerStatus;
+import java.util.Set;
+import lombok.Builder;
+import lombok.extern.slf4j.Slf4j;
+
+@Builder
+@Slf4j
+public record PlayersStatusDTO(
+        Long roomId,
+        Set<PlayerStatus> playerStatuses
+) {
+
+}
