+++ chatty-be/src/main/java/com/chatty/chatty/game/controller/dto/model/MakeQuizRequest.java
@@ -1,0 +1,19 @@
+package com.chatty.chatty.game.controller.dto.model;
+
+import java.util.List;
+import lombok.Builder;
+
+@Builder
+public record MakeQuizRequest(
+
+        Long roomId,
+
+        Integer numOfQuiz,
+
+        String type,
+
+        List<String> files
+
+) {
+
+}
