+++ chatty-be/src/main/java/com/chatty/chatty/game/controller/dto/model/MakeQuizResponse.java
@@ -1,0 +1,12 @@
+package com.chatty.chatty.game.controller.dto.model;
+
+import lombok.Builder;
+
+@Builder
+public record MakeQuizResponse(
+
+        Long roomId,
+
+        Long quizDocId
+) {
+}
