--- chatty-be/src/main/java/com/chatty/chatty/model/config/RestClientConfig.java
@@ -1,16 +1,0 @@
-package com.chatty.chatty.model.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestClient;
-
-@Configuration
-public class RestClientConfig {
-
-    @Bean
-    public RestClient restClient() {
-        return RestClient.builder()
-                .baseUrl("http://localhost:8000")
-                .build();
-    }
-}
