--- chatty-be/src/main/java/com/chatty/chatty/model/controller/dto/MakeQuizRequest.java
@@ -1,19 +1,0 @@
-package com.chatty.chatty.model.controller.dto;
-
-import java.util.List;
-import lombok.Builder;
-
-@Builder
-public record MakeQuizRequest(
-
-        Long roomId,
-
-        Integer numOfQuiz,
-
-        String type,
-
-        List<String> files
-
-) {
-
-}
