--- chatty-be/src/main/java/com/chatty/chatty/model/controller/dto/MakeQuizResponse.java
@@ -1,12 +1,0 @@
-package com.chatty.chatty.model.controller.dto;
-
-import lombok.Builder;
-
-@Builder
-public record MakeQuizResponse(
-
-        Long roomId,
-
-        Long quizDocId
-) {
-}
