--- chatty-be/src/main/java/com/chatty/chatty/quizroom/controller/dto/ChatRequest.java
@@ -1,11 +1,0 @@
-package com.chatty.chatty.quizroom.controller.dto;
-
-public record ChatRequest(
-        ChatType chatType,
-
-        Long roomId,
-
-        String message
-) {
-
-}
