--- chatty-be/src/main/java/com/chatty/chatty/quizroom/controller/dto/ChatResponse.java
@@ -1,19 +1,0 @@
-package com.chatty.chatty.quizroom.controller.dto;
-
-import java.time.LocalDateTime;
-import lombok.Builder;
-
-@Builder
-public record ChatResponse(
-        ChatType chatType,
-
-        Long roomId,
-
-        Long userId,
-
-        String message,
-
-        LocalDateTime time
-) {
-
-}
