--- chatty-be/src/main/java/com/chatty/chatty/quizroom/controller/dto/ChatType.java
@@ -1,7 +1,0 @@
-package com.chatty.chatty.quizroom.controller.dto;
-
-public enum ChatType {
-
-    TEXT,
-    EMOJI
-}
