--- chatty-be/src/main/java/com/chatty/chatty/quizroom/domain/entity/Status.java
@@ -1,8 +1,0 @@
-package com.chatty.chatty.quizroom.domain.entity;
-
-public enum Status {
-
-    READY,
-    STARTED,
-    FINISHED
-}
