+++ chatty-be/src/main/java/com/chatty/chatty/quizroom/entity/Status.java
@@ -1,0 +1,8 @@
+package com.chatty.chatty.quizroom.entity;
+
+public enum Status {
+
+    READY,
+    STARTED,
+    FINISHED
+}
