+++ chatty-be/src/main/java/com/chatty/chatty/quizroom/repository/QuizRoomRepository.java
@@ -1,6 +1,6 @@
 package com.chatty.chatty.quizroom.repository;
 
-import com.chatty.chatty.quizroom.domain.entity.QuizRoom;
+import com.chatty.chatty.quizroom.entity.QuizRoom;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
